+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Documentation of release versions of `nacc-form-validator`
 ## 0.2.0
 
 * Added documentation for macOS install and build
-* Renamed `validator` to `nacc_form_validator` for a more unique namespace
+* Renamed `validator` to `nacc_form_validator` for a more unique namespace, and imported `NACCValidator` into `__init__.py` for easier access
 
 ## 0.1.1
